refactor: orginize components
