+++ apps/client/src/components/dispatch/active-incidents/active-incidents.tsx
@@ -10,12 +10,12 @@ import { useActiveIncidentsTable } from "hooks/realtime/use-active-incidents-tab
 import { AlertModal } from "components/modal/AlertModal";
 import useFetch from "lib/useFetch";
 import type { IncidentInvolvedUnit, LeoIncident } from "@snailycad/types";
-import { InvolvedUnitsColumn } from "./active-incidents/InvolvedUnitsColumn";
+import { InvolvedUnitsColumn } from "./columns/involved-units-column";
 import { DndActions } from "types/dnd-actions";
 import { classNames } from "lib/classNames";
 import { useDispatchState } from "state/dispatch/dispatch-state";
 import type { PostIncidentsData, PutIncidentByIdData } from "@snailycad/types/api";
-import { CallDescription } from "./active-calls/CallDescription";
+import { CallDescription } from "../active-calls/CallDescription";
 
 import dynamic from "next/dynamic";
 import { useActiveIncidents } from "hooks/realtime/useActiveIncidents";
