+++ apps/client/src/components/dispatch/active-units/officers/columns/active-call-column.tsx
@@ -11,7 +11,7 @@ import { classNames } from "lib/classNames";
 
 const Manage911CallModal = dynamic(
   async () =>
-    (await import("components/dispatch/modals/manage-911-call/manage-911-call-modal"))
+    (await import("components/dispatch/active-calls/modals/manage-911-call-modal"))
       .Manage911CallModal,
   { ssr: false },
 );
