+++ apps/client/src/components/leo/ModalButtons.tsx
@@ -7,7 +7,7 @@ import useFetch from "lib/useFetch";
 import { makeUnitName } from "lib/utils";
 import { isUnitCombined, isUnitOfficer } from "@snailycad/utils";
 import * as modalButtons from "components/modal-buttons/buttons";
-import { ModalButton } from "components/modal-buttons/ModalButton";
+import { ModalButton } from "components/modal-buttons/modal-button";
 import type { PostLeoTogglePanicButtonData } from "@snailycad/types/api";
 import { useActiveDispatchers } from "hooks/realtime/use-active-dispatchers";
 import { ModalIds } from "types/modal-ids";
@@ -19,8 +19,8 @@ import { usePermission } from "hooks/usePermission";
 import { defaultPermissions } from "@snailycad/permissions";
 import dynamic from "next/dynamic";
 import { ImageWrapper } from "components/shared/image-wrapper";
-import { ActiveCallColumn } from "components/dispatch/active-units/officers/active-call-column";
-import { ActiveIncidentColumn } from "components/dispatch/active-units/officers/active-incident-column";
+import { ActiveCallColumn } from "components/dispatch/active-units/officers/columns/active-call-column";
+import { ActiveIncidentColumn } from "components/dispatch/active-units/officers/columns/active-incident-column";
 import { PrivateMessagesButton } from "./private-messages/private-messages-button";
 import { PrivateMessagesModal } from "components/dispatch/active-units/private-messages/private-messages-modal";
